Add new HUD assets and associated metadata files
- Added multiple PNG images for HUD elements including 1.png, 2.png, 3.png, and various skillbox images.
- Included metadata files for each new image to define texture import settings.
- Created a new fonts meta file to manage font assets.

+++ Assets/Scripts/PlayerReliquia.cs
@@ -53,8 +53,8 @@ public class PlayerReliquia : MonoBehaviour
             case TipoBonusReliquia.VidaMaxima:
                 if (stats != null)
                 {
-                    stats.coracoesMaximos += (int)reliquia.valor;
-                    stats.coracoesAtuais += (int)reliquia.valor;
+                    stats.coracoesMaximos += reliquia.valor;
+                    stats.coracoesAtuais += reliquia.valor;
                 }
                 break;
 
@@ -82,7 +82,7 @@ public class PlayerReliquia : MonoBehaviour
             case TipoBonusReliquia.VidaMaxima:
                 if (stats != null)
                 {
-                    stats.coracoesMaximos -= (int)reliquia.valor;
+                    stats.coracoesMaximos -= reliquia.valor;
                     stats.coracoesAtuais = Mathf.Min(stats.coracoesAtuais, stats.coracoesMaximos);
                 }
                 break;
